fix(docs): Standardize PHPDoc headers for src and test classes and update related documentation.
