+++ src/Exception/Message.php
@@ -7,24 +7,9 @@ namespace UIAwesome\Html\Core\Exception;
 use function sprintf;
 
 /**
- * Represents error message templates.
+ * Represents error message templates for attribute exceptions.
  *
- * This enum defines formatted error messages for various error conditions that may occur during operations such as
- * tag handling and event management.
- *
- * It provides message templates that can be formatted at call sites.
- *
- * Each case represents a specific type of error, with a message template that can be populated with dynamic values
- * using the {@see Message::getMessage()} method.
- *
- * Each message template can be formatted with arguments.
- *
- * Key features.
- * - Can be used by exception call sites that need formatted messages.
- * - Defines message templates as enum cases.
- * - Formats templates with `sprintf()` via {@see Message::getMessage()}.
- * - Supports message formatting with dynamic parameters.
- * - Uses the enum case `value` as the template string.
+ * Use {@see Message::getMessage()} to format the template with `sprintf()` arguments.
  *
  * @copyright Copyright (C) 2025 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
@@ -69,14 +54,17 @@ enum Message: string
     /**
      * Returns the formatted message string for the error case.
      *
+     * Usage example:
+     * ```php
+     * throw new \InvalidArgumentException(
+     *     \UIAwesome\Html\Core\Exception\Message::CANNOT_INSTANTIATE_ABSTRACT_CLASS
+     *         ->getMessage(\App\Html\MyAbstractClass::class),
+     * );
+     * ```
+     *
      * @param int|string ...$argument Values to insert into the message template.
      *
      * @return string Formatted error message with interpolated arguments.
-     *
-     * Usage example:
-     * ```php
-     * throw new InvalidArgumentException(Message::CANNOT_INSTANTIATE_ABSTRACT_CLASS->getMessage(MyAbstractClass::class));
-     * ```
      */
     public function getMessage(int|string ...$argument): string
     {
