+++ src/Provider/ThemeProviderInterface.php
@@ -7,17 +7,21 @@ namespace UIAwesome\Html\Core\Provider;
 use UIAwesome\Html\Core\Base\BaseTag;
 
 /**
- * Interface for applying theme configurations to tag instances.
+ * Defines theme providers for tag instances.
  *
- * Defines a contract for classes that return theme-based configuration arrays for {@see BaseTag} objects. The returned
- * definitions are applied by {@see \UIAwesome\Html\Core\Base\BaseTag::addThemeProvider()}.
+ * Usage example:
+ * ```php
+ * use UIAwesome\Html\Core\Base\BaseTag;
+ * use UIAwesome\Html\Core\Provider\ThemeProviderInterface;
  *
- * Intended for use in theme provider implementations that map a theme identifier to configuration definitions.
- *
- * Key features.
- * - Centralizes theme application logic for tag instances.
- * - Maps a theme identifier to configuration arrays.
- * - Returns configuration arrays consumed by {@see \UIAwesome\Html\Core\Factory\SimpleFactory::configure()}.
+ * final class ButtonThemeProvider implements ThemeProviderInterface
+ * {
+ *     public function apply(BaseTag $tag, string $theme): array
+ *     {
+ *         return $theme === 'dark' ? ['class' => 'btn-dark'] : [];
+ *     }
+ * }
+ * ```
  *
  * @copyright Copyright (C) 2025 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
@@ -25,10 +29,18 @@ use UIAwesome\Html\Core\Base\BaseTag;
 interface ThemeProviderInterface
 {
     /**
-     * Applies theme configuration to the given tag instance.
+     * Returns theme configuration for a tag instance.
      *
-     * Returns an associative array of configuration values for the specified tag and theme, supporting extensible and
-     * consistent theme application across the system.
+     * Usage example:
+     * ```php
+     * public function apply(\UIAwesome\Html\Core\Base\BaseTag $tag, string $theme): array
+     * {
+     *     return match ($theme) {
+     *         'dark' => ['class' => 'btn-dark'],
+     *         default => [],
+     *     };
+     * }
+     * ```
      *
      * @param BaseTag $tag Tag instance to which the theme is applied.
      * @param string $theme Theme identifier to apply.
@@ -36,22 +48,6 @@ interface ThemeProviderInterface
      * @return array Associative array of theme configuration values.
      *
      * @phpstan-return mixed[]
-     *
-     * Usage example:
-     * ```php
-     * public function apply(BaseTag $tag, string $theme): array
-     * {
-     *     if ($tag instanceof ButtonTag === false) {
-     *        return [];
-     *     }
-     *
-     *     return match ($theme) {
-     *         'dark' => ['class' => 'btn-dark'],
-     *         'light' => ['class' => 'btn-light'],
-     *         default => [],
-     *     };
-     * }
-     * ```
      */
     public function apply(BaseTag $tag, string $theme): array;
 }
